Use per-component Suspense for Connections Instant Nav
Paint the real header chrome outside Suspense so the hero animates once,
and stream only the connections list behind its own skeleton. Warm
per-user cache still fills the list without a skeleton flash.
