+++ app/[locale]/dashboard/connections/components/connections-page-skeleton.tsx
@@ -1,38 +1,24 @@
 import { Skeleton } from '@/components/ui/skeleton'
 
 /**
- * Instant navigation shell for /dashboard/connections.
- * Layout-only placeholders (no real copy) so the hero entrance animation
- * runs once when the real page mounts. Warm `"use cache"` skips this shell.
+ * Per-component Suspense fallback for the connections list only.
+ * Page chrome (title, actions) lives outside this boundary and paints instantly.
  */
-export function ConnectionsPageSkeleton() {
+export function ConnectionsListSkeleton() {
   return (
-    <div
-      className="min-h-[calc(100vh-var(--navbar-height,4rem))] bg-[oklch(0.97_0_0)] text-[oklch(0.17_0_0)] dark:bg-[oklch(0.17_0_0)] dark:text-[oklch(0.93_0_0)]"
-      aria-busy="true"
-      aria-live="polite"
-    >
-      <div className="mx-auto max-w-[1440px] px-5 py-10 sm:px-8 md:py-14 lg:px-12">
-        <header className="mb-12 md:mb-16">
-          <Skeleton className="h-9 w-56 rounded-sm bg-black/10 md:h-12 md:w-72 dark:bg-white/10" />
-          <Skeleton className="mt-3 h-5 w-full max-w-xl rounded-sm bg-black/10 dark:bg-white/10" />
-          <Skeleton className="mt-2 h-5 w-2/3 max-w-md rounded-sm bg-black/10 dark:bg-white/10" />
-          <div className="mt-8 flex flex-wrap items-center gap-3">
-            <Skeleton className="h-11 w-40 rounded-sm bg-black/10 dark:bg-white/10" />
-            <Skeleton className="h-11 w-28 rounded-sm bg-black/10 dark:bg-white/10" />
-          </div>
-        </header>
-
-        <div className="space-y-14 md:space-y-16">
-          {[0, 1, 2].map((section) => (
-            <section key={section} className="space-y-3">
-              <Skeleton className="h-7 w-40 rounded-sm bg-black/10 dark:bg-white/10" />
-              <Skeleton className="h-24 w-full rounded-sm bg-black/5 dark:bg-white/5" />
-              <Skeleton className="h-24 w-full rounded-sm bg-black/5 dark:bg-white/5" />
-            </section>
-          ))}
-        </div>
-      </div>
+    <div className="space-y-14 md:space-y-16" aria-busy="true" aria-live="polite">
+      {[0, 1, 2].map((section) => (
+        <section key={section} className="space-y-3">
+          <Skeleton className="h-7 w-40 rounded-sm bg-black/10 dark:bg-white/10" />
+          <Skeleton className="h-24 w-full rounded-sm bg-black/5 dark:bg-white/5" />
+          <Skeleton className="h-24 w-full rounded-sm bg-black/5 dark:bg-white/5" />
+        </section>
+      ))}
     </div>
   )
+}
+
+/** @deprecated Prefer ConnectionsListSkeleton + ConnectionsPageChrome */
+export function ConnectionsPageSkeleton() {
+  return <ConnectionsListSkeleton />
 }
