+++ app/[locale]/dashboard/connections/loading.tsx
@@ -1,5 +1,10 @@
-import { ConnectionsPageSkeleton } from './components/connections-page-skeleton'
+import { ConnectionsPageChrome } from './components/connections-page-chrome'
+import { ConnectionsListSkeleton } from './components/connections-page-skeleton'
 
 export default function ConnectionsLoading() {
-  return <ConnectionsPageSkeleton />
+  return (
+    <ConnectionsPageChrome>
+      <ConnectionsListSkeleton />
+    </ConnectionsPageChrome>
+  )
 }
